+++ modules/services/analytics/src/main/java/au/com/shiftyjelly/pocketcasts/analytics/AnalyticsEvent.kt
@@ -1,0 +1,81 @@
+package au.com.shiftyjelly.pocketcasts.analytics
+
+enum class AnalyticsEvent(val key: String) {
+
+    /* App Lifecycle */
+    APPLICATION_INSTALLED("application_installed"),
+    APPLICATION_UPDATED("application_updated"),
+    APPLICATION_OPENED("application_opened"),
+    APPLICATION_CLOSED("application_closed"),
+
+    /* User lifecycle events */
+    USER_SIGNED_IN("user_signed_in"),
+    USER_SIGNIN_FAILED("user_signin_failed"),
+    USER_ACCOUNT_DELETED("user_account_deleted"),
+    USER_PASSWORD_UPDATED("user_password_updated"),
+    USER_EMAIL_UPDATED("user_email_updated"),
+    USER_PASSWORD_RESET("user_password_reset"),
+    USER_ACCOUNT_CREATED("user_account_created"),
+    USER_ACCOUNT_CREATION_FAILED("user_account_creation_failed"),
+    USER_SIGNED_OUT("user_signed_out"),
+
+    /* Plus Upsell */
+    PLUS_PROMOTION_SHOWN("plus_promotion_shown"),
+    PLUS_PROMOTION_DISMISSED("plus_promotion_dismissed"),
+    PLUS_PROMOTION_UPGRADE_BUTTON_TAPPED("plus_promotion_upgrade_button_tapped"),
+
+    /* Setup Account */
+    SETUP_ACCOUNT_SHOWN("setup_account_shown"),
+    SETUP_ACCOUNT_DISMISSED("setup_account_dismissed"),
+    SETUP_ACCOUNT_BUTTON_TAPPED("setup_account_button_tapped"),
+
+    /* Sign in */
+    SIGNIN_SHOWN("signin_shown"),
+    SIGNIN_DISMISSED("signin_dismissed"),
+
+    /* Select Account Type */
+    SELECT_ACCOUNT_TYPE_SHOWN("select_account_type_shown"),
+    SELECT_ACCOUNT_TYPE_DISMISSED("select_account_type_dismissed"),
+    SELECT_ACCOUNT_TYPE_BUTTON_TAPPED("select_account_type_button_tapped"),
+
+    /* Create Account */
+    CREATE_ACCOUNT_SHOWN("create_account_shown"),
+    CREATE_ACCOUNT_DISMISSED("create_account_dismissed"),
+    CREATE_ACCOUNT_NEXT_BUTTON_TAPPED("create_account_next_button_tapped"),
+
+    /* Terms of Use */
+    TERMS_OF_USE_SHOWN("terms_of_use_shown"),
+    TERMS_OF_USE_DISMISSED("terms_of_use_dismissed"),
+    TERMS_OF_USE_ACCEPTED("terms_of_use_accepted"),
+    TERMS_OF_USE_REJECTED("terms_of_use_rejected"),
+
+    /* Select Payment Frequency */
+    SELECT_PAYMENT_FREQUENCY_SHOWN("select_payment_frequency_shown"),
+    SELECT_PAYMENT_FREQUENCY_DISMISSED("select_payment_frequency_dismissed"),
+    SELECT_PAYMENT_FREQUENCY_NEXT_BUTTON_TAPPED("select_payment_frequency_next_button_tapped"),
+
+    /* Confirm Payment */
+    CONFIRM_PAYMENT_SHOWN("confirm_payment_shown"),
+    CONFIRM_PAYMENT_DISMISSED("confirm_payment_dismissed"),
+    CONFIRM_PAYMENT_CONFIRM_BUTTON_TAPPED("confirm_payment_confirm_button_tapped"),
+
+    /* Purchase */
+    PURCHASE_SUCCESSFUL("purchase_successful"),
+    PURCHASE_CANCELLED("purchase_cancelled"),
+    PURCHASE_FAILED("purchase_failed"),
+
+    /* Newsletter Opt In */
+    NEWSLETTER_OPT_IN_CHANGED("newsletter_opt_in_changed"),
+
+    /* Forgot Password */
+    FORGOT_PASSWORD_SHOWN("forgot_password_shown"),
+    FORGOT_PASSWORD_DISMISSED("forgot_password_dismissed"),
+
+    /* Account Updated */
+    ACCOUNT_UPDATED_SHOWN("account_updated_shown"),
+    ACCOUNT_UPDATED_DISMISSED("account_updated_dismissed"),
+
+    /* Account Details */
+    ACCOUNT_DETAILS_SHOW_PRIVACY_POLICY("account_details_show_privacy_policy"),
+    ACCOUNT_DETAILS_SHOW_TOS("account_details_show_tos"),
+}