+++ modules/services/analytics/src/main/java/au/com/shiftyjelly/pocketcasts/analytics/AnalyticsTrackerWrapper.kt
@@ -1,0 +1,21 @@
+package au.com.shiftyjelly.pocketcasts.analytics
+
+import javax.inject.Inject
+
+class AnalyticsTrackerWrapper @Inject constructor() {
+    fun track(event: AnalyticsEvent, properties: Map<String, Any> = emptyMap()) {
+        AnalyticsTracker.track(event, properties)
+    }
+
+    fun refreshMetadata() {
+        AnalyticsTracker.refreshMetadata()
+    }
+
+    fun flush() {
+        AnalyticsTracker.flush()
+    }
+
+    fun clearAllData() {
+        AnalyticsTracker.clearAllData()
+    }
+}