+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/type/SubscriptionFrequency.kt
@@ -1,7 +1,8 @@
 package au.com.shiftyjelly.pocketcasts.models.type
 
-enum class SubscriptionFrequency {
-    NONE,
-    MONTHLY,
-    YEARLY
+enum class SubscriptionFrequency(val label: String) {
+    NONE("none"),
+    MONTHLY("monthly"),
+    YEARLY("yearly");
+    override fun toString() = label
 }