+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/type/SubscriptionType.kt
@@ -1,6 +1,7 @@
 package au.com.shiftyjelly.pocketcasts.models.type
 
-enum class SubscriptionType {
-    NONE,
-    PLUS
+enum class SubscriptionType(val label: String) {
+    NONE("none"),
+    PLUS("plus");
+    override fun toString() = label
 }