+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/ServerCallback.kt
@@ -1,8 +1,5 @@
 package au.com.shiftyjelly.pocketcasts.servers
 
-interface ServerCallback<T> {
-
-    fun callFailed(errorCode: Int, userMessage: String?, userMessageId: Int?, serverMessage: String?, throwable: Throwable?)
-
+interface ServerCallback<T> : ServerFailure {
     fun dataReturned(result: T?)
 }