+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release notes
 
+7.22.0
+-----
+
+*   Bug Fixes:
+    *   Fix playing on Chromecast always shows buffering.
+        ([#254](https://github.com/Automattic/pocket-casts-android/pull/254)).
+    *   Fix Plus subscription confirm button not always working.
+        ([#284](https://github.com/Automattic/pocket-casts-android/pull/284)).
+
+
 ### 7.21.1
 
 *   Bug Fixes: